Make RenderSoruce and OpenXRViewConfig into OpenXRViewSource
