+++ app/src/main/cpp/openxr-view-source.cc
@@ -2,11 +2,11 @@
 
 #include "logger.h"
 #include "openxr-util.h"
-#include "openxr-view-config.h"
+#include "openxr-view-source.h"
 
 namespace zen::display_system::oculus {
 
-OpenXRViewConfig::~OpenXRViewConfig()
+OpenXRViewSource::~OpenXRViewSource()
 {
   for (auto swapchain : swapchains_) {
     xrDestroySwapchain(swapchain.handle);
@@ -14,7 +14,7 @@ OpenXRViewConfig::~OpenXRViewConfig()
 }
 
 bool
-OpenXRViewConfig::Init()
+OpenXRViewSource::Init()
 {
   XrSystemProperties system_properties{XR_TYPE_SYSTEM_PROPERTIES};
   IF_XR_FAILED (err, xrGetSystemProperties(context_->instance(),
@@ -134,7 +134,7 @@ OpenXRViewConfig::Init()
         config_view.recommendedSwapchainSampleCount;
     swapchain_create_info.usageFlags = XR_SWAPCHAIN_USAGE_COLOR_ATTACHMENT_BIT;
 
-    OpenXRViewConfig::Swapchain swapchain;
+    OpenXRViewSource::Swapchain swapchain;
     swapchain.width = config_view.recommendedImageRectWidth;
     swapchain.height = config_view.recommendedImageRectHeight;
     IF_XR_FAILED (err, xrCreateSwapchain(context_->session(),
@@ -168,6 +168,44 @@ OpenXRViewConfig::Init()
   views_.resize(view_count, {XR_TYPE_VIEW});
 
   return true;
+}
+
+void
+OpenXRViewSource::Process()
+{
+  if (context_->is_session_running() == false) return;
+
+  XrFrameWaitInfo frame_wait_info{XR_TYPE_FRAME_WAIT_INFO};
+  XrFrameState frame_state{XR_TYPE_FRAME_STATE};
+  IF_XR_FAILED (err,
+      xrWaitFrame(context_->session(), &frame_wait_info, &frame_state)) {
+    LOG_ERROR("%s", err.c_str());
+    loop_->Terminate();
+    return;
+  }
+
+  XrFrameBeginInfo frame_begin_info{XR_TYPE_FRAME_BEGIN_INFO};
+  IF_XR_FAILED (err, xrBeginFrame(context_->session(), &frame_begin_info)) {
+    LOG_ERROR("%s", err.c_str());
+    loop_->Terminate();
+    return;
+  }
+
+  std::vector<XrCompositionLayerBaseHeader *> layers;
+  if (frame_state.shouldRender == XR_TRUE) {
+    // TODO: render
+  }
+
+  XrFrameEndInfo frame_end_info{XR_TYPE_FRAME_END_INFO};
+  frame_end_info.displayTime = frame_state.predictedDisplayTime;
+  frame_end_info.environmentBlendMode = context_->environment_blend_mode();
+  frame_end_info.layerCount = (uint32_t)layers.size();
+  frame_end_info.layers = layers.data();
+  IF_XR_FAILED (err, xrEndFrame(context_->session(), &frame_end_info)) {
+    LOG_ERROR("%s", err.c_str());
+    loop_->Terminate();
+    return;
+  }
 }
 
 }  // namespace zen::display_system::oculus
